Rename module directory in ibcratelimit.
