+++ app/app.go
@@ -142,8 +142,8 @@ import (
 	ibchookskeeper "github.com/provenance-io/provenance/x/ibchooks/keeper"
 	ibchookstypes "github.com/provenance-io/provenance/x/ibchooks/types"
 	"github.com/provenance-io/provenance/x/ibcratelimit"
-	"github.com/provenance-io/provenance/x/ibcratelimit/ibcratelimitmodule"
 	ibcratelimitkeeper "github.com/provenance-io/provenance/x/ibcratelimit/keeper"
+	ibcratelimitmodule "github.com/provenance-io/provenance/x/ibcratelimit/module"
 	ibcratelimittypes "github.com/provenance-io/provenance/x/ibcratelimit/types"
 	"github.com/provenance-io/provenance/x/marker"
 	markerkeeper "github.com/provenance-io/provenance/x/marker/keeper"
